It works wuhuuuu
it was a connection string issue

+++ src/Chirp.Razor/Pages/UserTimeline.cshtml.cs
@@ -26,13 +26,13 @@ public class UserTimelineModel : PageModel
     [FromQuery(Name = "page")]
     public int? pageNum { get; set; }
 
-  /*   [FromQuery(Name ="follow")]
+    [FromQuery(Name ="follow")]
     public int? follow{ get; set; }
 
-    [FromQuery(Name ="unfollow")] */
+    [FromQuery(Name ="unfollow")]
     public int? unfollow{ get; set; }
 
-    public ActionResult OnGet(string author)
+    public async Task<ActionResult> OnGet(string author)
     {
         if (pageNum.HasValue)
         {
@@ -43,17 +43,17 @@ public class UserTimelineModel : PageModel
             Cheeps = _service.GetCheepsFromAuthor(author, pageNum);
         }
 
-        /* if (User.Identity?.IsAuthenticated == true  && User.Identity.Name != null) {
-            AuthorDTO currentUser = _service.GetAuthorByName(User.Identity.Name);
+        if (User.Identity?.IsAuthenticated == true  && User.Identity.Name != null) {
+            AuthorDTO currentUser = await _service.GetAuthorByName(User.Identity.Name);
             if (follow.HasValue && follow != null) 
             {
-                _service.AddFollowee(currentUser.AuthorId, (int)follow);
+                await _service.AddFollowee(currentUser.AuthorId, (int)follow);
             } 
             else if (unfollow.HasValue && unfollow != null) 
             {
-                _service.RemoveFollowee(currentUser.AuthorId, (int)unfollow);
+                await _service.RemoveFollowee(currentUser.AuthorId, (int)unfollow);
             }
-        } */ 
+        } 
 
         return Page();
     }
@@ -81,14 +81,14 @@ public class UserTimelineModel : PageModel
         return Redirect("/");
     }
 
-/*     public bool DoesFollow(int AuthorId) 
+    public async Task<bool> DoesFollow(int AuthorId) 
     {
         AuthorDTO? author = null;
         // Needs to be refactored into the get method so we does not call it 32 times per page load
         if (User?.Identity?.IsAuthenticated == true && User?.Identity?.Name != null) {
             
             if (User.Identity.Name != null) {
-                author = _service.GetAuthorByName(User.Identity.Name);
+                author = await _service.GetAuthorByName(User.Identity.Name);
             }
             
             if (author != null && author.Followed != null) {
@@ -100,5 +100,5 @@ public class UserTimelineModel : PageModel
             }
         }
         return false;
-    } */
+    }
 }
